server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.45 
psycopg2-binary==2.9.11
python-dotenv==1.2.1
google-generativeai==0.8.3
requests==2.32.3
aiofiles
python-multipart